+++ src/main/java/io/github/scaredsmods/scaredsfactions/api/client/menu/AbstractConfirmMenu.java
@@ -1,0 +1,26 @@
+package io.github.scaredsmods.scaredsfactions.api.client.menu;
+
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.ItemStack;
+import org.jetbrains.annotations.Nullable;
+
+public abstract class AbstractConfirmMenu extends AbstractContainerMenu {
+
+    public AbstractConfirmMenu(@Nullable MenuType<?> pMenuType, int pContainerId) {
+        super(pMenuType, pContainerId);
+    }
+
+    @Override
+    public ItemStack quickMoveStack(Player pPlayer, int pIndex) {
+        return ItemStack.EMPTY;
+    }
+
+    @Override
+    public boolean stillValid(Player pPlayer) {
+        return true;
+    }
+}