+++ src/main/java/io/github/scaredsmods/scaredsfactions/client/screen/ConfirmTransferOwnershipScreen.java
@@ -17,6 +17,7 @@
 package io.github.scaredsmods.scaredsfactions.client.screen;
 
 import com.mojang.blaze3d.systems.RenderSystem;
+import io.github.scaredsmods.scaredsfactions.api.client.screen.AbstractConfirmScreen;
 import io.github.scaredsmods.scaredsfactions.common.ScaredsFactionMod;
 import io.github.scaredsmods.scaredsfactions.client.screen.menu.ConfirmTransferOwnershipMenu;
 import io.github.scaredsmods.scaredsfactions.server.network.ModNetworks;
@@ -35,7 +36,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 
-public class ConfirmTransferOwnershipScreen extends AbstractContainerScreen<ConfirmTransferOwnershipMenu> {
+public class ConfirmTransferOwnershipScreen extends AbstractConfirmScreen<ConfirmTransferOwnershipMenu> {
 
 	private static final ResourceLocation TEXTURE = ScaredsFactionMod.id("textures/gui/container/confirm_transfer.png");
 	public Screen parent;
@@ -67,32 +68,6 @@ public class ConfirmTransferOwnershipScreen extends AbstractContainerScreen<Conf
 
 		this.addRenderableWidget(this.back);
 		this.addRenderableWidget(this.confirm);
-	}
-
-	public void resize(Minecraft pMinecraft, int pWidth, int pHeight) {
-		this.init(pMinecraft, pWidth, pHeight);
-	}
-
-	@Override
-	protected void renderBg(GuiGraphics pGuiGraphics, float pPartialTick, int pMouseX, int pMouseY) {
-		RenderSystem.setShader(GameRenderer::getPositionTexShader);
-		RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-		RenderSystem.setShaderTexture(0, TEXTURE);
-
-		pGuiGraphics.blit(TEXTURE, leftPos ,topPos, 0, 0, imageWidth, imageHeight);
-	}
-
-	@Override
-	public void render(GuiGraphics pGuiGraphics, int pMouseX, int pMouseY, float pPartialTick) {
-		renderBackground(pGuiGraphics);
-		super.render(pGuiGraphics, pMouseX, pMouseY, pPartialTick);
-		renderTooltip(pGuiGraphics, pMouseX, pMouseY);
-	}
-
-	@Override
-	public void onClose() {
-		if (this.minecraft == null) return;
-		this.minecraft.setScreen(this.parent);
 	}
 
 }