+++ src/main/java/io/github/scaredsmods/scaredsfactions/common/ScaredsFactionMod.java
@@ -51,8 +51,8 @@ public class ScaredsFactionMod
 		return ResourceLocation.fromNamespaceAndPath(MOD_ID, name);
 	}
 
-	public static String permission(String name) {
-		return "scaredsfactions." + name;
+	public static String permission(String perm) {
+		return String.format("scaredsfactions.%s", perm);
 	}